add policy

+++ app/Http/Controllers/ThreadsController.php
@@ -118,13 +118,18 @@ class ThreadsController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
+        // if($thread->user_id != auth()->id()){
+        //     abort(403, 'You do not have permission to do this.');
+        // }
+
+        $this->authorize('update', $thread);
+        
         $thread->replies()->delete();
         $thread->delete();
-
-        if(request()->wantsJson){
+        
+        if(request()->wantsJson()){
             return response([], 204);
         }
-        
         return redirect('/threads');
     }
 }
